chore: update embedding service URL
