--- backend/AzurePhotoFlow.EmbeddingService/Controllers/EmbeddingGeneratorController.cs
@@ -1,25 +1,0 @@
-using Api.Interfaces;
-using Microsoft.AspNetCore.Mvc;
-
-namespace AzurePhotoFlow.Services;
-
-[ApiController]
-[Route("api/[controller]")]
-public class EmbeddingGeneratorController : ControllerBase
-{
-    private readonly IEmbeddingGeneratorService _generatorService;
-
-    public EmbeddingGeneratorController(IEmbeddingGeneratorService generatorService)
-    {
-        _generatorService = generatorService;
-    }
-
-    [HttpPost("generate")]
-    public async Task<IActionResult> Generate([FromBody] EmbeddingRequest request)
-    {
-        await _generatorService.GenerateAsync(request.ProjectName, request.DirectoryName, request.Timestamp);
-        return Ok();
-    }
-}
-
-public record EmbeddingRequest(string ProjectName, string DirectoryName, DateTime Timestamp);
