+++ backend/AzurePhotoFlow.EmbeddingService/Interfaces/IEmbeddingService.cs
@@ -1,6 +1,6 @@
 namespace Api.Interfaces;
 
-public interface IEmbeddingGeneratorService
+public interface IEmbeddingService
 {
     Task GenerateAsync(string projectName, string directoryName, DateTime timestamp);
 }
