+++ docker-compose.yml
@@ -37,7 +37,7 @@ services:
       - MINIO_ENDPOINT=${MINIO_ENDPOINT}
       - MINIO_ACCESS_KEY=${MINIO_ACCESS_KEY}
       - MINIO_SECRET_KEY=${MINIO_SECRET_KEY}
-      - EMBEDDING_SERVICE_URL=http://embedding:80
+      - EMBEDDING_SERVICE_URL=http://embedding:80/api/Embedding
     networks:
       - app-network
